@@ -42,11 +42,12 @@
     "electron": "^7.0.1",
     "electron-builder": "^22.3.2",
     "electron-builder-notarize": "^1.1.2",
-    "eslint": "^5.6.1",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-react": "^7.11.1",
+    "eslint": "^5.16.0 || ^6.1.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react-hooks": "^1.7.0",
     "svg2png": "^4.1.1"
   },
   "dependencies": {
@@ -68,10 +69,11 @@
     "mousetrap": "^1.6.1",
     "p-queue": "^6.2.0",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-hammerjs": "^1.0.1",
     "react-sortable-hoc": "^1.5.3",
+    "react-use": "^13.24.0",
     "read-chunk": "^2.0.0",
     "string-to-stream": "^1.1.1",
     "strong-data-uri": "^1.0.5",
