@@ -133,7 +133,7 @@ async function cutMultiple({
   const outFiles = [];
 
   let i = 0;
-  // eslint-disable-next-line no-restricted-syntax
+  // eslint-disable-next-line no-restricted-syntax,no-unused-vars
   for (const { cutFrom, cutTo, cutToApparent } of segments) {
     const ext = path.extname(filePath) || `.${format}`;
     const cutSpecification = `${formatDuration(cutFrom, true)}-${formatDuration(cutToApparent, true)}`;
